requirements.txt
pandas
openpyxl
numpy
neuralforecast
matplotlib
python-dotenv
scikit-learn
pymssql
